+++ docs/changelog.md
@@ -8,6 +8,7 @@ Sprint 8 / March 14, 2020
 - turned off max-statements rule in ESLint config.
 - exported type definition for Button component to be used in Timer component.
 - refactored Timer component's stopTimer and startTimer functions into one toggleTimer function.
+- memoized Button child components inside Timer component to prevent unnecessary rerenders.
 
 ## Release 0.7.0
 Sprint 7 / February 29, 2020